+++ pytool/__init__.py
@@ -14,7 +14,7 @@
     limitations under the License.
 
 """
-__version__ = '2.3.2'
+__version__ = '2.4.0-dev'
 
 
 from pytool import (